+++ examples/consumer/main.go
@@ -4,7 +4,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/go-redis/redis"
+	"github.com/go-redis/redis/v7"
 	"github.com/kak-tus/ami"
 )