+++ examples/producer/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"time"
 
-	"github.com/go-redis/redis"
+	"github.com/go-redis/redis/v7"
 	"github.com/kak-tus/ami"
 )